Note model classes moved to separate package
